Get indicators from specified source in stock_db

Generic function to get indicator timeseries from specified source in stock_db.

get_indicators_from_source(
  stock_db,
  indicator_source,
  indicator_codes = NULL,
  ouput_format = c("long", "wide"),
  ...
)

# S4 method for gta_db
get_indicators_from_source(
  stock_db,
  indicator_source,
  indicator_codes,
  ouput_format,
  ...,
  data_filters,
  date_fields,
  retain_fields
)

Arguments

stock_db
A stock database object to operate.

indicator_source
A name of table or file(with extension) in which indicators are stored. For file, formats of rds,csv are supported.

indicator_codes
A vector of indicator code. Default NULL return all indicators in the source.

ouput_format
Output format for data frame, i.e "long" and "wide":

in long format: code of indicator is stored in "ind_code", value of indicator is stored in "ind_value";

in wide format: value of indicator is stored in field named with the name of indicator.

...
Extra arguments to be passed to methods.

data_filters
A list of filter to apply on result data. The list consists of items which has field as its name and filter expr string as its content, default is list( typrep = "typrep == 'A'", markettype = "markettype == 21"). Only valid for gta_db.

date_fields
A character vector of possible date field in original data, which will be renamed as "date". default is c("date", "accper", "trddt", "trdmnt", "trdynt","clsdt", "shrchdgt"). Only valid for gta_db.

retain_fields
A character vector of possible fields to retain in final dataset, default is c("stkcd", "indcd"). Only valid for gta_db.

Value

A data frame of indicator timeseries if succeed, otherwise NULL.

Functions

get_indicators_from_source,gta_db-method: get indicator timeseries from specified source in a database of gta_db class

See also