Get a dataset of a list of stock_cd from a table in stock_db

Generic function to get a dataset of a list of stock_cd from table in stock.

get_stock_dataset(stock_db, table_name, stock_cd_list = NULL, ...)

# S4 method for gta_db
get_stock_dataset(stock_db, table_name, stock_cd_list = NULL, ...)

Arguments

stock_db
A stock database object to operate.

table_name
Name of target table.

stock_cd_list
A character vector of stock cd, default value of NULL means. all stock data will be returned.

...
Extra arguments to be passed to methods.

Value

A data frame on succeed otherwise NULL.

Functions

get_stock_dataset,gta_db-method: get a dataset of a list of stock_cd from table in a database of gta_db class

See also

Examples

if (FALSE) {
ds_trd_dalyr.df <- get_stock_dataset(stock_db,
                            table_name = "TRD_Dalyr_鏃ヤ釜鑲″洖鎶ョ巼",
                            stock_cd_list = c("600066", "000550"))
}