Get stocks excess return from stocks return and riskfree_rate

Get excess return timeseries by combining stocks return and riskfree_rate

stocks_excess_return(
  ts_stocks_return,
  ts_riskfree_rate,
  period = c("day", "month", "quarter", "year"),
  period_date = c("start", "end")
)

Arguments

ts_stocks_return
A timeseries of stocks_return

ts_riskfree_rate
A timeseries of riskfree_rate

period
A character of period, e.g. "day", "month", "quarter", "year". Default "day".

period_date
A character of period_date format, e.g. "start", "end", "start" format date as start of the period, "end" format date as end of period. Default "start".

Value

A timeseries of excess return of stocks.