Compute a lagged version of timeseries

Generic function to shift data of timeseries(except date) forward and backward according current timeline

ts_lag(ts_dataset, k = 1, trim = TRUE, ...)

# S3 method for tbl_df
ts_lag(
  ts_dataset,
  k = 1,
  trim = TRUE,
  ...,
  date_index_field = c("date"),
  key_fields = NULL,
  parallel = TRUE
)

# S3 method for timeSeries
ts_lag(ts_dataset, k = 1, trim = TRUE, ...)

Arguments

ts_dataset
a timeseries of tibble/timeseries.

k
an integer value. The number of lags (in units of observations). By default 1.

k > 0: shift forward, using earlier data as current data

k = 0: don't shift, keeping original data

k < 0: shift backward, using later data as current data

see details for more info.

trim
A logical value. By default TRUE, the first missing observation in the return series will be removed.

...
Arguments passed to other methods

date_index_field
Name of date index field of ts_df for resample, default 'date', Column must be date-like. Only be used for tibble dataset.

key_fields
A character vector of key fields, which identify unique observation in each date. Only be used for tibble dataset.

parallel
A logic to determine whether to use parallel processing. default TRUE means to use parallel processing.

Value

       A lagged timeseres

Details

There are two types of lag operation:

shift forward: mostly known as lag, i.e. move data to next k periods, which means we use earlier data as current data while keeping current timeline

shift backward: mostly known as head, i.e. move previous k periods, which means use later data as current data while keeping current timeline

Methods (by class)

tbl_df: Compute a lagged version of timeseries for tibble dataset

timeSeries: Compute a lagged version of timeSeries for timeSeries dataset

See also