Utils functions to control parallel process

Control parallel process for parallel computation.

enable_parallel(env_globals = .pkg_globals)

disable_parallel(env_globals = .pkg_globals)

parallel_status(env_globals = .pkg_globals)

Arguments

env_globals
a environment with parallel process information. Default .pkg_globals means to use global vars of the package. NULL means to use environment of caller.

Value

parallel_status return a list containing parallel status, see details.

Details

There are two levels of controlling parallel process for computation:

Global level: enable/disable parallel process for R session. Through enable_parallel, we could set parallel process ready for current R session by initiating back-end clusters, which normally are processes running on multiple cores in a computer. we could shut parallel process by using disable_parallel.

Function level: enable/disable parallel process for running functions. In a R session enabling parallel process, we could control a function being able to running parallel process, for example:

        compute_indicator(ts_compute_vars, compute_fun, ...,
                  date_index_field = "stkcd"
                  c("date"), key_fields = NULL,
                  parallel = getOption("zstmodelr.common.parallel", TRUE))

Its argument parallel controls whether the function run in parallel process or sequential process, whose default value is from options of zstmodelr.common.parallel or TRUE.

In addition, parallel_status could provide current status of parallel process and configuration options, which return a list including following fields:

cluster: a object with class of c("SOCKcluster", "cluster") containing information about back-end clusters for parallel process.

parallel_log: a path of parallel log file containing log info from parallel process.

foreach_workers: a number of current process workers ready for foreach computation.

zstmodelr.common.clusters: a number of clusters to be used for parallel process.

zstmodelr.common.parallel: a logical as default value for parallel argument of function with ability of parallel computation.

Functions

enable_parallel: enable parallel process by initiating back-end clusters.

disable_parallel: disable parallel process by stopping back-end clusters.

parallel_status: get current status of parallel process.

Examples

if (FALSE) {

# Set clusters number of parallel process
options(zstmodelr.common.clusters = 4)

# Enable parallel process for the R session
enable_parallel()

# Inspect status of parallel process
parallel_status()

# Conduct parallel computation ...

# compute indicators by parallel process
ds_indicator <- compute_indicator(ds_vars,
       compute_fun = ind_def_fun,
       date_index_field = "date",
       key_fields = "stkcd")

# compute indicators by sequential process
 ds_indicator <- compute_indicator(ds_vars,
       compute_fun = ind_def_fun,
       date_index_field = "date",
       key_fields = "stkcd",
       parallel = FALSE)

# Disable parallel process for the R session
disable_parallel()

}