zstmodelr: Stock Factor Models for China Stock Market Investment

An implementation of analytical tools of stock factors models for investing China stock market.

Details

The main goals of zstmodelr is to leverage traditional financial theory and modern ML method to build stock models for China stock market.

To learn more about zstmodelr, start with the vignettes: browseVignettes(package = "zstmodelr")

Options

testthat.use_colours: Should the output be coloured? (Default: TRUE).

testthat.summary.max_reports: The maximum number of detailed test reports printed for the summary reporter (default: 10).

testthat.summary.omit_dots: Omit progress dots in the summary reporter (default: FALSE).

See also

Author

Maintainer: Chris Zheng chriszheng@vip.sina.com.cn (ORCID)

Examples

if (FALSE) {
library(testthat)
a <- 9
expect_that(a, is_less_than(10))
expect_lt(a, 10)
}